Use a more standard way for git-ignoring a folder
The previous approach was working as expected in the git-related
features of VSCode, but wasn't working when using git from the terminal
(why? Are the git features in VSCode not just leveraging core git logic
to determine things like "what files are modified and not tracked?" That
feels surprising, but appears to be the case).

This new approach is simpler and works in both of those places. Yeah!