Add Makefile targets for building doom.wasm
The resultant WebAssembly module will still be further tweaked, but this
is a first pass, at least getting (I think) all the Doom code compiling
down into a single WebAssembly module.

Note that the WASI SDK (https://github.com/WebAssembly/wasi-sdk) was a
lot of help in this process, and plays a major role in how this
WebAssembly module is built.