Explicitly describe the interface to `doom.wasm``
by adding to this repo a text file that lists out all the imports and
exports in `doom.wasm`, the Doom WebAssembly module.

Do this by providing a Node.js script that can load and describe the
imports and exports of a WebAssembly module (Node.js was used due to the
ease at which it can load and interrogate WebAssembly modules), and by
providing a Makefile target that uses this Node.js script to generate
the file `doom.wasm.interface.txt` from `doom.wasm`.