Actually compile all Doom code into WASM module
Export the two functions that are used from the outside to drive the
game: doomgeneric_Create and doomgeneric_Tick (exported as 'initGame'
and 'tickGame').

This prevents the linker from stripping most of the game logic from the
module, because it's no longer dead code, and results in the the full
game's code actually existing in the module. Hence the size of the
resultant doom.wasm module is now 393k instead of 13k in size, which
seems a lot more reasonable for the size of the entire game (e.g. the
SDL-based build of the game is 641k in size on my system).

Also, instruct the linker to create an import for any missing functions.
So now the imports that are present on doom.wasm paint a clear picture
of what system-level features (a lot of them part of WASI) are needed to
run the game.