Switch to using the “reactor” executable model
See https://clang.llvm.org/docs/ClangCommandLineReference.html#cmdoption-clang-mexec-model

This change communicates "this WebAssembly module isn't a
command/program with a 'main' function that drives its execution,
instead this is a library with a lifetime controlled by the user".

This is where there is no longer an exported `_start()` function, and no
longer an explicit need for getting command line arguments from the
system.

Instead there is now an exported `_initialize()` function that the user
of doom.wasm is expected to call before they use any other function
exported by this module.