Remove logic related to error reporting via GUI
The removed approach sometimes involves spawning a system GUI to report
an error. Instead we're just going to rely on the error being reported
to stderr for now.

This change has been driven by the need to remove this code's dependency
on a 'system' function (which doesn't map to any WASI features) which
was being used to spawn a 'zenity' process in order to display errors
via a GUI.