Have wasm imports be better organized and named
Instead of blindly using the function names provided by doomgeneric,
instead augment the names to provide the right amount of context and be
better organized.

This involves conceptually splitting the imports between a 'ui' module
and a 'runtimeControl' module, which think will improve usability for
users of the wasm module.