Improve organization via better folder names
Do two things: (1) move all existing source code folders into a new
`doomgeneric` folder, and (2) rename the `target_SDL` folder to
`example_implementation`.

This will give us the structure we need to maintain and enhance a clear
interface named `doomgeneric` while also building on top of it in order
to build doom.wasm.

Also, the SDL implementation of the `doomgeneric` interface is only
unique in that it's an example, non-WebAssembly implementation. So we're
putting it in a folder with a name that better reflects this generic
need.