Have `gamemission` be determined by IWAD contents
This is being done because, in the future, WAD contents will be provided
solely through the `doomgeneric` interface instead of via command-line
arguments. And initial support for this will not involve the name of the
IWAD being provided through the `doomgeneric` interface.

It's not clear yet what features of obscure IWADs this may break, but
evaluating that is low priority for now, and will be addressed later.