Have game load all game data through DG_GetWads()
Instead of finding and loading WAD data through the filesystem, and
directing this process via command-line args, the game now loads all WAD
data through the new `doomgeneric` function `DG_GetWads()`.

The SDL implementation of this game still relies on command-line args to
be told which WAD files on disk to load, but the WebAssembly
implementation will now be free to provide raw WAD data directly to the
game and not rely on the filesystem.