Provide an easy way to run SDL Doom with a PWAD
As a way to test features related to loading PWADs, provide the make
target `run-with-test-pwad` which loads up SDL Doom with the PWAD
`GOTHIC2.WAD`.

This PWAD, and a Doom2-compatible IWAD, freedoom2.wad, have been added
to support this need.

And, because the fight this time and previously with the default
`check-added-large-files` pre-commit hook has been a pain, and this hook
will probably bring little benefit, it's being removed.