Implement game saving/loading through doomgeneric
Introduce new game saving/loading abstractions in `doomgeneric` that
only deal with save slot numbers as identifiers and raw streams of bytes
as the data. Leverage these abstractions in all code related to game
saving and loading, removing the need for core Doom to rely on a file
system for game saving/loading.

Also, implement these abstractions in the example SDL implementation by
reading and writing save game data to the filesystem, as was previously
done with Doom.