Simplify M_SetConfigDir based on how its called
There's only one caller of M_SetConfigDir, and it calls the function
with NULL, so we can just implement M_SetConfigDir take no args and
behave as if it was called with NULL, resulting in no logical change.

NOTE: this change is part of a larger effort to reduce the game's
dependency on storying configuration details on disk, in an effort to
remove this code's dependency on the WASI function
path_create_directory, and this commit accomplishes that detail with the
removal of a call to M_MakeDirectory.