Statically set the value of configdir
The assigned value of configdir is not dynamic, so we can safely assign
it statically and remove the use of M_SetConfigDir().

This change also removes a printf statement that advertised which
directory was being used "for configuration and saves". Game saves are
not necessarily being saved to disk anymore (this detail depends upon
the implementation of a few `doomgeneric` functions), and I don't
believe configuration details were being persisted at all. So the
removal of this printf statement is welcomed.