Support WAD loading via standard command line args
Expect the arg "-iwad <path>" to be used to specify which IWAD file to
load, and allow the use of "-file [<path> ...]" to specify paths to any
additional IWADs to load.

Support for this was added along with support for passing all
non-processed command line args through to Doom itself.