Add support for demo recording to `doomgeneric`
How the game responds to the player recording a demo is now handled via
a new `doomgeneric` function `DG_RecordDemo`.

The example/SDL implementation of DG_RecordDemo writes the demo data to
disk, just like was previously done before this commit.

Note: this change is purely motivated by the need to remove a call to
M_WriteFile, which depends upon WASI file system features that we're
aiming to not need.