Add support for screenshot taking to `doomgeneric`
How the game responds to the player taking a screenshot is now handled
via a new `doomgeneric` function `DG_PCXScreenshotTaken`.

The example/SDL implementation of DG_PCXScreenshotTaken writes the
screenshot data to disk, just like was previously done before this
commit.

Note: this change is purely motivated by the need to remove a call to
M_WriteFile, which depends upon WASI file system features that we're
aiming to not need.