Move onGameInit import to the 'loading' module
Because it's somewhat expected that the user will initialize the game ui
in their implementation of onGameInit, due to that function receiving
the width and height of the game's framebuffer, it initially made sense
to place this import in the 'ui' module.

But now I feel that this import is more directly related to the game
loading process, because it is called exactly once while the game is
loading, and generally reflects the one place where the user is given a
chance to respond to "game loaded", so this import has been moved to the
'loading' module.