Fix bug that was preventing any PWADs from loading
The implementation of DG_GetWads for the WeAssembly version of Doom was
never actually setting `numberOfPWads` to a non-default value on the
returned instance of DB_BytesForAllWads, so this was always `0` no matter
how many PWADs `getWadsSizes` was reporting. This issue was preventing
the successful load of any and all PWADs, but has now been fixed.