Remove all exports related to internal WASI impls
doom.wasm is supplying its own partial implementation of WASI by
exporting implementations of the WASI functions that are later hooked up
as imports for the actual WASI functions it needs.  Due to this approach
the resultant module exports many functions that aren't expected to be
actually used by the outside world.

Here we remove the exporting of these functions by using the Binaryen
tool `wasm-metadce`. This tool allows us to specify which of the
features exported by a module are actually used by the outside world,
and then it removes all exported features NOT used.

Note: the functions themselves are not removed from the module, only the
fact that the functions are being exported is removed.