Split up WebAssembly-specific code
Separate out the WebAssembly-specific code (currently just in
doom_wasm.c) into 'code that implements WASI features' and 'code that
does not'.

This allows us to more clearly split up the code that's related to the
interface that the WebAssembly module provides, and the code that's just
there to satisfy the WASI needs of this code (which also happens to
leverage the ability to export functions from a WebAssembly module).