Remove boring import 'ui.setWindowTitle'
It's unlikely that the window title that was being provided to this
import will be useful to any user of the WebAssembly module, since the
window title isn't terribly specific.

So, the cost of this import existing, forcing a user to provide an
implementation of it, is likely higher than the benefit of it existing.
Therefore, we're removing it.