Use clearly fixed-width types in imports/exports
Use a type like int32_t instead of int, for example, to make it clear to
the reader that the WebAssembly interface of Doom doesn't involve types
that may change their width based on the underlying platform.

Note: this doesn't actually change the WebAssembly interface of Doom, it
just makes the resultant interface easier to understand and doesn't rely
on the reader knowing that `int` is 32-bits wide on the WebAssembly
platform.