Use a supported integral type in reportKeyUp/Down
WebAssembly doesn't actually support 8-bit or unsigned integral types.
So even though we were using uint8_t as the type of the single argument
passed to reportKeyUp and reportKeyDown, the effective arg type of these
exported functions was `i32`.

This meant that the user of the module passing in a value outside the
uint8_t range would see no clear issue, and reportKeyUp and
reportKeyDown would (likely, this is implementation specific, I think)
just see the first byte of that value and be none the wiser about what
the user's actual behavior was.

By changing reportKeyUp and reportKeyDown to accept a int32_t these
functions now have the ability to see the full value the user passed in.
With no loss of information these functions are better handled to report
an error when the user has provided a value outside the uint8_t range.