Remove onExit because no current need for it
Currently the game just 'crashes' when the user wishes to 'exit' it, and
the underlying WASI function `proc_exit`` is never called.

This is not great, and should be addressed. But until that's addressed
there's no need to ask the user to provide an onExit import, which we'll
never actually use. So that import is being removed for now.