Fix unintended cast from uint32_t to int and back
The only users of I_GetTicks, which are I_GetTime and I_GetTimeMs, were
immediately casting I_GetTicks's return value to an uint32_t, which is
appropriate given that I_GetTicks itself casts an uint32_t to an int.

This intermediate use of `int` is confusing at best, and surely a
mistake at this point, so we're removing it.