Provide consistent default WAD loading behavior
WAD loading is implemented by a few WebAssembly imports filling in
needed data at a couple of spots in memory. By initializing all related
memory to reasonable defaults we get consistent behavior when an import
fails to fill in the requested values.

WAD loading still fails in that case, but at least we catch this case
and fail predictably instead of failing in a undetermined way.