Load Doom shareware WAD when no WAD is provided
Embed the contents of the Doom shareware WAD in the WebAssembly module
and load this WAD when the imported function `getWadsSizes` reports that
there are 0 WADs to load.

The behavior is helpful when the user of this module doesn't want to
fuss with manually providing a WAD to load and instead just wants the
Doom shareware WAD loaded. It's expected that this behavior will be
quite common.