Export only one function that does 'init' behavior
Instead of exporting both `_initialize()` and `initGame()`, and
expecting the user to call them in the correct order, we now export just
`initGame()` that performs the behavior of both of those function
at once. This provides a nicer user experience, reducing the barrier to
understanding doom.wasm's API.

Accomplishing this involves introducing another tiny utility WebAssembly
module and another use of `wasm-merge`, but that's pretty
straightforward.