Standardize .gitignore files
Make it so all directories are matched by something that's clearly a
directory. And have all matches be rooted at the root unless there's a
good reason why not.

Also, let all util/examples project completely manage ignored files on
their own so there's no confusion on which .gitignore is in charge of
what sections of this repo.